Delete and rename some icons
