--- love/src/components/GeneralPurpose/WindRose/WindRose.jsx
@@ -1,42 +1,0 @@
-import React, { Component } from 'react';
-// import PropTypes from 'prop-types';
-import styles from './WindRose.module.css';
-
-export default class WindRose extends Component {
-  render() {
-    return (
-      <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 56 66.5">
-        <path
-          className={styles.colorOne}
-          d="M56.12,38.35,38.7,34.83a10.83,10.83,0,0,0-5.11-5.11L30.07,12.3a1.09,1.09,0,0,0-2.14,0L24.41,29.72a10.83,10.83,0,0,0-5.11,5.11L1.88,38.35a1.09,1.09,0,0,0,0,2.14L19.3,44a10.85,10.85,0,0,0,5.11,5.1l3.52,17.42a1.09,1.09,0,0,0,2.14,0l3.52-17.42A10.85,10.85,0,0,0,38.7,44l17.42-3.53a1.09,1.09,0,0,0,0-2.14ZM37.71,42a1.08,1.08,0,0,0-.79.66,8.66,8.66,0,0,1-4.7,4.7,1.08,1.08,0,0,0-.66.79L29,60.81,26.44,48.13a1.06,1.06,0,0,0-.66-.79,8.61,8.61,0,0,1-4.69-4.7,1.12,1.12,0,0,0-.8-.66L7.61,39.42l12.68-2.56a1.09,1.09,0,0,0,.8-.66,8.56,8.56,0,0,1,4.69-4.69,1.09,1.09,0,0,0,.66-.8L29,18l2.56,12.67a1.12,1.12,0,0,0,.66.8,8.61,8.61,0,0,1,4.7,4.69,1.06,1.06,0,0,0,.79.66l12.68,2.56Z"
-          transform="translate(-1 -.92)"
-        />
-        <path
-          className={styles.colorOne}
-          d="M29,33.94a5.48,5.48,0,1,0,5.48,5.48A5.49,5.49,0,0,0,29,33.94Zm0,8.78a3.3,3.3,0,1,1,3.3-3.3A3.29,3.29,0,0,1,29,42.72Z"
-          transform="translate(-1 -.92)"
-        />
-        <path
-          className={styles.colorOne}
-          d="M10.31,34.73a1.13,1.13,0,0,0,.32.05,1.08,1.08,0,0,0,1-.77A18.13,18.13,0,0,1,23.59,22.1,1.09,1.09,0,0,0,22.94,20,20.32,20.32,0,0,0,9.59,33.36,1.09,1.09,0,0,0,10.31,34.73Z"
-          transform="translate(-1 -.92)"
-        />
-        <path
-          className={styles.colorOne}
-          d="M47.7,44.11a1.1,1.1,0,0,0-1.37.72A18.17,18.17,0,0,1,34.41,56.75a1.09,1.09,0,0,0-.72,1.37,1.1,1.1,0,0,0,1.05.76.93.93,0,0,0,.32-.05A20.26,20.26,0,0,0,48.41,45.48,1.1,1.1,0,0,0,47.7,44.11Z"
-          transform="translate(-1 -.92)"
-        />
-        <path
-          className={styles.colorOne}
-          d="M48.41,33.36A20.3,20.3,0,0,0,35.06,20a1.09,1.09,0,1,0-.65,2.09A18.13,18.13,0,0,1,46.33,34a1.08,1.08,0,0,0,1,.77,1.26,1.26,0,0,0,.33-.05A1.09,1.09,0,0,0,48.41,33.36Z"
-          transform="translate(-1 -.92)"
-        />
-        <path
-          className={styles.colorTwo}
-          d="M31,7.12,26.94,1.65h0V7.12h-.18V1.42H27l4.09,5.46h0V1.42h.18v5.7Z"
-          transform="translate(-1 -.92)"
-        />
-      </svg>
-    );
-  }
-}
