+++ love/src/components/icons/AddIcon/AddIcon.jsx
@@ -3,18 +3,12 @@ import styles from './AddIcon.module.css';
 
 function AddIcon(props) {
   return (
-    <svg viewBox="-1 -1 30 30" {...props}>
+    <svg viewBox="0 0 48.02 48.02" {...props}>
       <path
-        className={styles.path}
-        d="M13.9,0c-2.3,0-4.6,0.6-6.7,1.7C7,1.8,6.9,2.2,7,2.4c0.1,0.3,0.5,0.4,0.7,0.2c1.9-1,4-1.6,6.1-1.6
-          c7.1,0,12.8,5.7,12.8,12.8S21,26.7,13.9,26.7S1.1,21,1.1,13.9c0-3.4,1.4-6.7,3.8-9.1c0.2-0.2,0.2-0.6,0-0.8C4.7,3.8,4.4,3.8,4.1,4
-          C1.5,6.6,0,10.2,0,13.9c0,7.7,6.2,13.9,13.9,13.9s13.9-6.2,13.9-13.9S21.6,0,13.9,0z"
-      />
-      <path
-        className={styles.path}
-        d="M6.4,14.8c-0.3,0-0.6-0.3-0.6-0.6s0.2-0.6,0.6-0.6h6.8V6.9c0-0.3,0.3-0.6,0.6-0.6s0.6,0.3,0.6,0.6v6.8H21
-          c0.3,0,0.6,0.2,0.6,0.6s-0.3,0.6-0.6,0.6h-6.8v6.8c0,0.3-0.2,0.6-0.6,0.6s-0.6-0.3-0.6-0.6v-6.8H6.4z"
-      />
+        d="M24 .47a24 24 0 1024 24 24 24 0 00-24-24zm0 7.68A2.88 2.88 0 0126.84 11v10.59H37.4a2.88 2.88 0 110 5.76H26.84v10.57a2.89 2.89 0 01-5.77 0V27.35H10.51a2.88 2.88 0 110-5.76h10.56V11A2.88 2.88 0 0124 8.15z"
+        className={styles["cls-1"]}
+        transform="translate(.05 -.47)"
+      ></path>
     </svg>
   );
 }
