--- love/src/components/icons/ContactIcon/ContactIcon.jsx
@@ -1,26 +1,0 @@
-import React from 'react';
-import styles from './ContactIcon.module.css';
-
-export default class ContactIcon extends React.Component {
-  render() {
-    return (
-      <svg className={styles.contactIcon} viewBox="0 0 31.9 39.2">
-        <path
-          d="M17.5 19.2s.1 0 0 0c.1 0 .1 0 0 0h.1c1.1 0 2-.4 2.7-1.2 1.5-1.7 1.2-4.5 1.2-4.8-.1-2-1.1-3-1.8-3.5-.6-.3-1.3-.5-2-.5-.4 0-1.3.1-2 .5-.8.5-1.8 1.4-1.9 3.5 0 .3-.3 3.1 1.2 4.8.5.8 1.4 1.2 2.5 1.2zm-2.8-5.8s0-.1 0 0c.1-2.8 2.1-3.1 2.9-3.1 1 0 2.8.4 2.9 3 0 0 .3 2.6-.9 4-.5.5-1.1.8-2 .8-.8 0-1.5-.3-2-.8-1.2-1.3-.9-3.9-.9-3.9z"
-          className={styles.st0}
-        ></path>
-        <path
-          d="M25.4 23.9v-.1c0-.8-.1-2.5-1.7-3.1-1.7-.4-3.1-1.4-3.2-1.4-.2-.2-.6-.1-.7.1-.2.2-.1.6.1.7.1 0 1.6 1.1 3.5 1.6.9.3 1 1.3 1 2.1v.1c0 .3 0 .9-.1 1.2-.6.4-3 1.6-6.7 1.6s-6.1-1.2-6.7-1.6c-.1-.3-.1-.8-.1-1.2v-.1c0-.9.1-1.8 1-2.1 1.9-.5 3.4-1.5 3.5-1.6.2-.2.3-.5.1-.7-.2-.2-.5-.3-.7-.1 0 0-1.4 1-3.2 1.4-1.7.6-1.7 2.3-1.7 3.1v.1c0 .2 0 1.2.2 1.7 0 .1.1.2.2.2.1.1 2.9 1.8 7.4 1.8s7.3-1.7 7.4-1.8c.1-.1.2-.1.2-.2.2-.5.2-1.5.2-1.7z"
-          className={styles.st0}
-        ></path>
-        <path
-          d="M5.2 30.8V6.5c0-1 .8-1.9 1.9-1.9h21.1c1 0 1.9.8 1.9 1.9v24.3c0 1-.8 1.9-1.9 1.9H7c-1 0-1.8-.8-1.8-1.9z"
-          className={styles.st1}
-        ></path>
-        <path d="M1 18.7L5.2 18.7" className={styles.st1}></path>
-        <path d="M1 10.1L5.2 10.1" className={styles.st1}></path>
-        <path d="M1 27L5.2 27" className={styles.st1}></path>
-      </svg>
-    );
-  }
-}
