+++ love/src/components/icons/DebugIcon/DebugIcon.jsx
@@ -3,22 +3,17 @@ import styles from './DebugIcon.module.css';
 
 function DebugIcon(props) {
   return (
-    <svg viewBox="0 0 28.9 26.8" {...props}>
+    <svg viewBox="0 0 48.27 55.46" {...props}>
       <path
-        className={styles.path}
-        d="M9,5.9c0,0.3,0.3,0.6,0.6,0.6s0.6-0.3,0.6-0.6c0-2.2,1.8-4,4-4s4,1.8,4,4c0,0.3,0.3,0.6,0.6,0.6
-          s0.6-0.3,0.6-0.6c0-2.9-2.4-5.3-5.3-5.3S9,3,9,5.9z"
-      />
+        d="M12.05 2.16l5.15 8.78a9.07 9.07 0 00-1.55 5.17 20.79 20.79 0 018.59-1.9 21.41 21.41 0 018.59 1.9v-.17a8.59 8.59 0 00-1.54-5l5.15-8.78L33.52.44l-4.81 8.09a7.92 7.92 0 00-4.47-1.2 9.56 9.56 0 00-4.46 1.2L15 .44z"
+        className={styles["cls-1"]}
+        transform="translate(-.02 -.44)"
+      ></path>
       <path
-        className={styles.path}
-        d="M27.8,15.6h-5.1V13h2c1.6,0,2.8-1.3,2.8-2.8V7.1c0-0.3-0.3-0.6-0.6-0.6c-0.3,0-0.6,0.3-0.6,0.6v3.2
-          c0,0.9-0.7,1.6-1.6,1.6h-2v-1.5c0-0.9-0.7-1.6-1.6-1.6H7.4c-0.9,0-1.6,0.7-1.6,1.6v1.5h-2c-0.9,0-1.6-0.7-1.6-1.6V7.1
-          c0-0.3-0.3-0.6-0.6-0.6S0.9,6.7,0.9,7.1v3.2c0,1.6,1.3,2.8,2.8,2.8h2v2.6H0.6c-0.3,0-0.6,0.3-0.6,0.6c0,0.3,0.3,0.6,0.6,0.6h5.1
-          v2.6h-2c-1.6,0-2.8,1.3-2.8,2.8v3.2c0,0.3,0.3,0.6,0.6,0.6s0.6-0.3,0.6-0.6v-3.2c0-0.9,0.7-1.6,1.6-1.6h2.1
-          c0.5,3.2,3.3,5.7,6.7,5.7h3.3c3.4,0,6.2-2.5,6.7-5.7h2.1c0.9,0,1.6,0.7,1.6,1.6v3.2c0,0.3,0.3,0.6,0.6,0.6s0.6-0.3,0.6-0.6v-3.2
-          c0-1.6-1.3-2.8-2.8-2.8h-2v-2.6h5.1c0.3,0,0.6-0.3,0.6-0.6C28.5,15.9,28.2,15.6,27.8,15.6z M7.1,19.7v-9.4c0-0.2,0.2-0.4,0.4-0.4
-          h6.2v15.3h-1C9.5,25.2,7.1,22.7,7.1,19.7z M15.9,25.2h-1V9.9h6.2c0.2,0,0.4,0.2,0.4,0.4v9.4C21.4,22.7,19,25.2,15.9,25.2z"
-      />
+        d="M48.29 43.5v-8.79c-6.35-3.27-4.81-2.41-7.38-3.79a14.39 14.39 0 00-1.72-4.48c1.89-1 .51-.34 7.21-3.61v-8.62H43v6.55c-5 2.58-3.78 1.89-5.84 2.93a16.81 16.81 0 00-13.06-6 17.37 17.37 0 00-13 6L9.64 23l-4.29-2.07v-6.72H1.91V23l5.67 2.93 1.55.69a19.51 19.51 0 00-1.72 4.48C5 32.3 6.38 31.61 0 34.88v8.79h3.46v-6.55l3.43-1.72a15.73 15.73 0 002.92 9.3c-1.54.86-.68.35-4.64 2.41v8.79h3.44v-6.55c2.23-1.2 1.55-.69 3.44-1.72a17 17 0 0024.05 0c1.89 1 1.2.52 3.43 1.72v6.55H43v-8.79c-3.61-1.89-2.75-1.37-4.64-2.41a17.46 17.46 0 002.92-9.3l3.44 1.72v6.55h3.6z"
+        className={styles["cls-1"]}
+        transform="translate(-.02 -.44)"
+      ></path>
     </svg>
   );
 }
