--- love/src/components/icons/NewAddIcon/NewAddIcon.module.css
@@ -1,7 +1,0 @@
-.path {
-  fill: var(--secondary-font-color);
-  stroke: var(--secondary-font-color);
-}
-.cls-1 {
-  fill:#798d99;
-}
