--- love/src/components/icons/NewRedoIcon/NewRedoIcon.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import styles from './NewRedoIcon.module.css';
-
-function NewRedoIcon(props) {
-  return (
-    <svg viewBox="0 0 46.36 49.01" {...props}>
-      <path
-        d="M.53 26.81a20.71 20.71 0 0020.52 23.41h13.5a1.72 1.72 0 001.7-1.71v-5.3a1.72 1.72 0 00-1.7-1.71H20.94A12 12 0 019.19 27a12.24 12.24 0 0112.11-9.52h6.29v5.88a1.7 1.7 0 002.63 1.39l15.72-10.2a1.69 1.69 0 000-2.84L30.22 1.5a1.69 1.69 0 00-2.63 1.4v5.87h-5.87C11.2 8.77 1.82 16.35.53 26.81z"
-        className={styles["cls-1"]}
-        transform="translate(-.36 -1.21)"
-      ></path>
-    </svg>
-  );
-}
-
-export default NewRedoIcon;
