--- love/src/components/icons/NewSaveIcon/NewSaveIcon.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import styles from './NewSaveIcon.module.css';
-
-function NewSaveIcon(props) {
-  return (
-    <svg viewBox="0 0 55.34 55.34" x="0px" y="0px" className={styles["svg"]}>
-     <path className={styles["cls-1"]} d="M16.89,1.19V15h0a1.38,1.38,0,0,0,1.38,1.39H37.64A1.38,1.38,0,0,0,39,15V1.19h0A1.39,1.39,0,0,1,40.41-.2h1.85a5.58,5.58,0,0,1,3.92,1.62L54,9.25h0a5.53,5.53,0,0,1,1.62,3.91V49.61h0a5.52,5.52,0,0,1-5.53,5.53H5.82A5.53,5.53,0,0,1,.28,49.61V5.34h0A5.54,5.54,0,0,1,5.82-.2H15.5A1.39,1.39,0,0,1,16.89,1.19ZM29.34-.2H32.1a1.39,1.39,0,0,1,1.39,1.39v8.3h0a1.38,1.38,0,0,1-1.39,1.38H29.34A1.38,1.38,0,0,1,28,9.49V1.19h0A1.39,1.39,0,0,1,29.34-.2Zm-11.07,36h0A9.69,9.69,0,0,0,28,45.46h0a9.69,9.69,0,0,0,9.69-9.69h0A9.69,9.69,0,0,0,28,26.09h0A9.69,9.69,0,0,0,18.27,35.77Z" transform="translate(-0.28 0.2)"/>
-    </svg>
-  );
-}
-
-export default NewSaveIcon;
