--- love/src/components/icons/NewThumbnailIcon/NewThumbnailIcon.module.css
@@ -1,11 +1,0 @@
-.active {
-  fill: var(--base-font-color);
-}
-
-.inactive {
-  fill: var(--secondary-font-color);
-}
-
-.cls-1 {
-  fill:#798f9b;
-}
