--- love/src/components/icons/NewUndoIcon/NewUndoIcon.module.css
@@ -1,8 +1,0 @@
-.path {
-  fill: var(--secondary-font-color);
-  stroke: var(--secondary-font-color);
-}
-
-.cls-1 { 
-  fill:#798d99;
-}
