+++ love/src/components/icons/RedoIcon/RedoIcon.jsx
@@ -3,14 +3,12 @@ import styles from './RedoIcon.module.css';
 
 function RedoIcon(props) {
   return (
-    <svg viewBox="0 0 27.8 27.8" {...props}>
+    <svg viewBox="0 0 46.36 49.01" {...props}>
       <path
-        className={styles.path}
-        d="M18,10.2c0.4,0,0.7-0.3,0.7-0.7l0-6.8l6.6,6.8l-6.6,6.8v-1.9c0-0.4-0.3-0.7-0.7-0.7c-0.4,0-0.7,0.3-0.7,0.7
-          l0,3.6c0,0.3,0.2,0.6,0.4,0.7c0.3,0.1,0.6,0.1,0.8-0.2l8.3-8.5c0.3-0.3,0.3-0.8,0-1l-8.3-8.5c-0.2-0.2-0.5-0.3-0.8-0.2
-          c-0.3,0.1-0.4,0.4-0.4,0.7l0,7.8h-0.4c0,0-3.1-0.2-5.1,0.2C4.4,10.3-0.8,17.3,0.1,24.5c0,0.4,0.3,0.6,0.6,0.6c0,0,0.1,0,0.1,0
-          c0.3-0.1,0.6-0.4,0.5-0.8C0.5,17.9,5.3,11.6,12,10.4C15.5,9.8,18,10.2,18,10.2z"
-      />
+        d="M.53 26.81a20.71 20.71 0 0020.52 23.41h13.5a1.72 1.72 0 001.7-1.71v-5.3a1.72 1.72 0 00-1.7-1.71H20.94A12 12 0 019.19 27a12.24 12.24 0 0112.11-9.52h6.29v5.88a1.7 1.7 0 002.63 1.39l15.72-10.2a1.69 1.69 0 000-2.84L30.22 1.5a1.69 1.69 0 00-2.63 1.4v5.87h-5.87C11.2 8.77 1.82 16.35.53 26.81z"
+        className={styles["cls-1"]}
+        transform="translate(-.36 -1.21)"
+      ></path>
     </svg>
   );
 }
