+++ love/src/components/icons/RedoIcon/RedoIcon.module.css
@@ -2,3 +2,7 @@
   fill: var(--secondary-font-color);
   stroke: var(--secondary-font-color);
 }
+
+.cls-1 {
+  fill:#798d99;
+}
