+++ love/src/components/icons/SaveIcon/SaveIcon.jsx
@@ -3,18 +3,8 @@ import styles from './SaveIcon.module.css';
 
 function SaveIcon(props) {
   return (
-    <svg viewBox="0 0 27 28" x="0px" y="0px" {...props}>
-      <path
-        className={styles.path}
-        d="M21.5,1.2c0.1,0,0.3,0.1,0.4,0.2l3.8,3.8c0.1,0.1,0.2,0.2,0.2,0.4v14.4c0,0.3,0.3,0.6,0.6,0.6l0,0
-        c0.3,0,0.6-0.3,0.6-0.6V5.2C27,5,27,4.9,26.9,4.8l-4.5-4.5c-0.1-0.1-0.2-0.2-0.4-0.2h-1.1H5.1H0.6C0.3,0.1,0,0.4,0,0.7v26.5
-        c0,0.3,0.3,0.6,0.6,0.6h3.4h18.6h3.9c0.3,0,0.6-0.3,0.6-0.6v-3.4c0-0.3-0.2-0.6-0.5-0.6l0,0c-0.3,0-0.6,0.2-0.6,0.6V26
-        c0,0.3-0.3,0.6-0.6,0.6h-2.3c-0.3,0-0.6-0.3-0.6-0.6V15.3c0-0.3-0.3-0.6-0.6-0.6H4.5c-0.3,0-0.6,0.3-0.6,0.6V26
-        c0,0.3-0.3,0.6-0.6,0.6H1.7c-0.3,0-0.6-0.3-0.6-0.6V1.8c0-0.3,0.3-0.6,0.6-0.6h3.4h1.1h13.5v7.3c0,0.3-0.3,0.6-0.6,0.6H6.8
-        c-0.3,0-0.6-0.3-0.6-0.6V3.4c0-0.3-0.2-0.5-0.5-0.5h0c-0.3,0-0.6,0.3-0.6,0.6v6.3c0,0.3,0.3,0.6,0.6,0.6h14.6
-        c0.3,0,0.6-0.3,0.6-0.6V1.2 M5.1,26v-9.6c0-0.3,0.3-0.6,0.6-0.6h15.2c0.3,0,0.6,0.3,0.6,0.6V26c0,0.3-0.3,0.6-0.6,0.6H5.6
-        C5.3,26.6,5.1,26.3,5.1,26z"
-      />
+    <svg viewBox="0 0 55.34 55.34" x="0px" y="0px" className={styles["svg"]}>
+     <path className={styles["cls-1"]} d="M16.89,1.19V15h0a1.38,1.38,0,0,0,1.38,1.39H37.64A1.38,1.38,0,0,0,39,15V1.19h0A1.39,1.39,0,0,1,40.41-.2h1.85a5.58,5.58,0,0,1,3.92,1.62L54,9.25h0a5.53,5.53,0,0,1,1.62,3.91V49.61h0a5.52,5.52,0,0,1-5.53,5.53H5.82A5.53,5.53,0,0,1,.28,49.61V5.34h0A5.54,5.54,0,0,1,5.82-.2H15.5A1.39,1.39,0,0,1,16.89,1.19ZM29.34-.2H32.1a1.39,1.39,0,0,1,1.39,1.39v8.3h0a1.38,1.38,0,0,1-1.39,1.38H29.34A1.38,1.38,0,0,1,28,9.49V1.19h0A1.39,1.39,0,0,1,29.34-.2Zm-11.07,36h0A9.69,9.69,0,0,0,28,45.46h0a9.69,9.69,0,0,0,9.69-9.69h0A9.69,9.69,0,0,0,28,26.09h0A9.69,9.69,0,0,0,18.27,35.77Z" transform="translate(-0.28 0.2)"/>
     </svg>
   );
 }
