+++ love/src/components/icons/SaveIcon/SaveIcon.module.css
@@ -1,4 +1,13 @@
-.path {
+/* .path {
   fill: var(--secondary-font-color);
   stroke: var(--secondary-font-color);
+} */
+.svg {
+  vertical-align: top;
+  fill: var(--secondary-font-color);
+  stroke: var(--base-font-color);
+}
+.cls-1 {
+  fill:#798d99;
+  fill-rule:evenodd;
 }
