+++ love/src/components/icons/ThumbnailIcon/ThumbnailIcon.module.css
@@ -5,3 +5,7 @@
 .inactive {
   fill: var(--secondary-font-color);
 }
+
+.cls-1 {
+  fill:#798f9b;
+}
