+++ love/src/components/icons/UndoIcon/UndoIcon.jsx
@@ -3,14 +3,12 @@ import styles from './UndoIcon.module.css';
 
 function UndoIcon(props) {
   return (
-    <svg viewBox="0 0 27 25.1" {...props}>
+    <svg viewBox="0 0 46.36 49.01" {...props}>
       <path
-        className={styles.path}
-        d="M9.1,10.2c-0.4,0-0.7-0.3-0.7-0.7l0-6.8L1.8,9.5l6.6,6.8v-1.9c0-0.4,0.3-0.7,0.7-0.7c0.4,0,0.7,0.3,0.7,0.7
-          l0,3.6c0,0.3-0.2,0.6-0.4,0.7c-0.3,0.1-0.6,0.1-0.8-0.2L0.2,10C0,9.7,0,9.2,0.2,8.9l8.3-8.5c0.2-0.2,0.5-0.3,0.8-0.2
-          c0.3,0.1,0.4,0.4,0.4,0.7l0,7.8h0.4c0,0,3.1-0.2,5.1,0.2c7.4,1.4,12.6,8.4,11.6,15.6c0,0.4-0.3,0.6-0.6,0.6c0,0-0.1,0-0.1,0
-          c-0.3-0.1-0.6-0.4-0.5-0.8c0.9-6.4-3.9-12.7-10.6-13.9C11.5,9.8,9.1,10.2,9.1,10.2z"
-      />
+        d="M46.55 26.81A20.72 20.72 0 0126 50.22H12.53a1.73 1.73 0 01-1.71-1.71v-5.3a1.73 1.73 0 011.71-1.71h13.6A12 12 0 0037.89 27a12.26 12.26 0 00-12.12-9.49h-6.29v5.88a1.69 1.69 0 01-2.62 1.39L1.13 14.55a1.69 1.69 0 010-2.84L16.86 1.5a1.68 1.68 0 012.62 1.4v5.87h5.88c10.52 0 19.9 7.58 21.19 18.04z"
+        className={styles["cls-1"]}
+        transform="translate(-.36 -1.21)"
+      ></path>
     </svg>
   );
 }
