+++ love/src/components/icons/WindRose/WindRose.module.css
